Move functionality to gcp.py
